refactor: facade 메서드 service로 이관, facade 제거
