+++ src/main/kotlin/com/devfloor/untitled/article/application/ArticleService.kt
@@ -4,8 +4,12 @@ import com.devfloor.untitled.article.domain.Article
 import com.devfloor.untitled.article.domain.ArticleRepository
 import com.devfloor.untitled.articlefavorite.application.ArticleFavoriteService
 import com.devfloor.untitled.articlehashtag.application.ArticleHashtagService
+import com.devfloor.untitled.articlehashtag.domain.ArticleHashtag
 import com.devfloor.untitled.articleoption.application.ArticleOptionService
 import com.devfloor.untitled.common.exception.EntityNotFoundException
+import com.devfloor.untitled.hashtag.application.HashtagService
+import com.devfloor.untitled.option.application.OptionService
+import com.devfloor.untitled.user.domain.User
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -16,31 +20,61 @@ class ArticleService(
     private val articleHashtagService: ArticleHashtagService,
     private val articleFavoriteService: ArticleFavoriteService,
     private val articleOptionService: ArticleOptionService,
+    private val hashtagService: HashtagService,
+    private val optionService: OptionService,
 ) {
     @Transactional(readOnly = true)
-    fun showByArticleId(articleId: Long): ArticleResponse {
-        repository.findByIdOrNull(articleId)
-            ?.let {
+    fun showByArticleId(articleId: Long): ArticleResponse =
+        showByArticleIdOrNull(articleId)
+            .let {
                 val articleOptions = articleOptionService.showAllByArticle(it)
                 val articleHashtags = articleHashtagService.showAllByArticle(it)
                 val articleFavorites = articleFavoriteService.showAllByArticle(it)
 
-                return ArticleResponse(
+                ArticleResponse(
                     articleOptions = articleOptions,
                     article = it,
                     articleHashtags = articleHashtags,
                     articleFavorites = articleFavorites,
                 )
             }
-        throw EntityNotFoundException("사용자가 요청한 리소스가 없습니다")
-    }
+
+    fun showByArticleIdOrNull(articleId: Long) = repository.findByIdOrNull(articleId)
+        ?: throw EntityNotFoundException("사용자가 요청한 리소스가 없습니다")
 
     fun showAll(): List<Article> = repository.findAll()
 
-    fun create(article: Article): Article = repository.save(article)
+    @Transactional
+    fun create(request: ArticleRequest, user: User): Long {
+        val article = repository.save(request.toArticle(user))
 
-    fun modify(id: Long, title: String?, content: String): Article =
-        showById(id).apply { modify(title, content) }
+        if (request.hasOptions) {
+            optionService.showAllByOptionType(request.options)
+                .let { articleOptionService.createAll(article, it) }
+        }
 
-    fun destroyById(id: Long) = repository.deleteById(id)
+        request.hashtags
+            .map { hashtagService.createByName(it) }
+            .map { articleHashtagService.create(ArticleHashtag(article, it)) }
+
+        return article.id
+    }
+
+    @Transactional
+    fun modifyByArticleId(
+        articleId: Long,
+        request: ArticleModifyRequest,
+    ) {
+        val article = showByArticleIdOrNull(articleId)
+            .apply { modify(request.title, request.content) }
+
+        request.hashtags
+            .map { hashtagService.createByName(it) }
+            .let { articleHashtagService.modifyByArticle(article, it) }
+
+        optionService.showAllByOptionType(request.options)
+            .let { articleOptionService.modifyByArticle(article, it) }
+    }
+
+    fun destroyById(articleId: Long) = repository.deleteById(articleId)
 }
